requirements.txt
bilibili-api-python>=16.1.1
aiohttp>=3.9.1
tqdm>=4.66.1 